move debounce to own file
